Update `parser-icu` example
