+++ examples/parser-icu/package.json
@@ -11,14 +11,14 @@
     "check:watch": "svelte-check --tsconfig ./tsconfig.json --watch"
   },
   "devDependencies": {
-    "@sveltejs/adapter-auto": "next",
-    "@sveltejs/kit": "next",
-    "svelte": "^3.44.0",
-    "svelte-check": "^2.2.6",
-    "svelte-preprocess": "^4.10.1",
+    "@sveltejs/adapter-auto": "^2.1.0",
+    "@sveltejs/kit": "^1.22.1",
+    "svelte": "^4.0.5",
+    "svelte-check": "^3.4.5",
+    "svelte-preprocess": "^5.0.4",
     "tslib": "^2.3.1",
-    "typescript": "^4.7.4",
-    "vite": "^3.0.2"
+    "typescript": "^5.1.6",
+    "vite": "^4.4.2"
   },
   "type": "module",
   "dependencies": {
