+++ examples/parser-icu/src/routes/+layout.svelte
@@ -1,0 +1,14 @@
+<script lang="ts">
+  import { t, locales, locale } from '$lib/translations';
+</script>
+
+<slot />
+<br />
+<br />
+<br />
+<br />
+<select bind:value="{$locale}">
+  {#each $locales as value}
+    <option value="{value}">{$t(`lang.${value}`)}</option>
+  {/each}
+</select>
