+++ examples/parser-icu/src/routes/+layout.ts
@@ -1,0 +1,8 @@
+import type { Load } from '@sveltejs/kit';
+import { loading } from '$lib/translations';
+
+export const load: Load = async () => {
+  await loading.toPromise();
+
+  return {};
+};
