--- examples/parser-icu/src/routes/__layout.svelte
@@ -1,21 +1,0 @@
-<script context="module" lang="ts">
-  import type { Load } from '@sveltejs/kit';
-  import { t, loading, locales, locale } from '$lib/translations';
-
-  export const load: Load = async () => {
-    await loading.toPromise();
-
-    return {};
-  }
-</script>
-
-<slot />
-<br />
-<br />
-<br />
-<br />
-<select bind:value="{$locale}">
-  {#each $locales as value}
-    <option value="{value}">{$t(`lang.${value}`)}</option>
-  {/each}
-</select>
